nearly done modularizing tasks
