+++ static/task.js
@@ -1,7 +1,9 @@
-if(typeof API_BASE == "undefined"){
-    console.log("api base not loaded")
-}
+import {getStatusColor} from './utilities.js';
+import { API_BASE } from './constants.js';
+import { formatDate } from './utilities.js';
+
 let currentRecordId = null;
+let independentTasks = []
 
 function openUpdateForm(id) {
     recordData = null;
@@ -22,7 +24,6 @@ function openUpdateForm(id) {
     
 }
 
-// Close update form
 function closeUpdateForm(event) {
     if (event && event.target !== event.currentTarget && !event.target.classList.contains('popup-close')) {
         return;
@@ -43,12 +44,6 @@ function removeToast(closeBtn) {
     }, 300);
 }
 
-// Close popup on Escape key
-document.addEventListener('keydown', function(e) {
-    if (e.key === 'Escape') {
-        closeUpdateForm();
-    }
-});
 
 function updateTask(){
     let task = {
@@ -69,4 +64,81 @@ function updateTask(){
     }).catch((e)=>{
         showPopup('error', 'Something failed:', `${e}`);
     })
+}
+
+export const getIndependentTasks = async()=>{
+    const task_response = await fetch(`${API_BASE}/independenttask`);
+    independentTasks = await task_response.json();
+}
+
+
+export const  renderIndependentTasksList=()=> {
+    const container = document.getElementById("independent-tasks-container");
+    container.innerHTML = independentTasks.map((task, index) => `
+        <div  onclick="openUpdateForm('${task.id}')" style="
+            background: #fff;
+            border: 1px solid #e1e5e9;
+            border-radius: 8px;
+            padding: 16px;
+            margin-bottom: 12px;
+            cursor: pointer;
+            transition: all 0.2s ease;
+            box-shadow: 0 1px 3px rgba(0, 0, 0, 0.1);
+        
+        " onmouseover="this.style.transform='translateY(-1px)'; this.style.boxShadow='0 4px 12px rgba(0, 0, 0, 0.15)'; this.style.borderColor='#007bff';" 
+           onmouseout="this.style.transform='translateY(0)'; this.style.boxShadow='0 1px 3px rgba(0, 0, 0, 0.1)'; this.style.borderColor='#e1e5e9';">
+            
+            <div style="display: flex;  justify-content: space-between; align-items: flex-start; margin-bottom: 8px;">
+                <h4 style="margin: 0; color: #2c3e50; font-size: 1.1rem; font-weight: 600;">
+                    ${task.name}
+                </h4>
+                <span class="status-badge" style="
+                    background: ${getStatusColor(task.status)};
+                    color: white;
+                    padding: 4px 8px;
+                    border-radius: 12px;
+                    font-size: 0.75rem;
+                    font-weight: 500;
+                    text-transform: capitalize;
+                ">
+                    ${task.status.replace('_', ' ')}
+                </span>
+            </div>
+            
+            <div style="display: flex; justify-content:space-between; align-items: center; gap: 16px; margin-bottom: 12px; color: #6c757d; font-size: 0.85rem;">
+                <div style="display: flex; align-items: center; gap: 4px;">
+                    <p>Start Date: </p>
+                    <span>${task.start_date ? formatDate(task.start_date) : 'No start'}</span>
+                </div>
+                <div style="display: flex; align-items: center; gap: 4px;">
+                    <p>End Date: </p>
+                    <span>${task.end_date ? formatDate(task.end_date) : 'No end'}</span>
+                </div>
+            </div>
+            
+            <p style="
+                margin: 0;
+                color: #495057;
+                line-height: 1.4;
+                font-size: 0.9rem;
+            ">
+                ${task.description || 'No description provided'}
+            </p>
+        </div>
+        
+    `).join('') || `
+        <div style="
+            text-align: center;
+            color: #6c757d;
+            padding: 3rem 1rem;
+            background: #f8f9fa;
+            border: 2px dashed #dee2e6;
+            border-radius: 8px;
+            margin: 1rem 0;
+        ">
+            <div style="font-size: 2rem; margin-bottom: 1rem;">📝</div>
+            <div style="font-size: 1.1rem; margin-bottom: 0.5rem;">No tasks added yet</div>
+            <div style="font-size: 0.9rem;">Add your first task to get started</div>
+        </div>
+    `;
 }
